requirements.txt


ipywidgets
openpyxl
pandas
pdf2image
pillow
poppler-utils
pymupdf
pytesseract
python-doctr
python-docx
python-dotenv
requests
streamlit
tqdm
unidecode
widgetsnbextension